Method to build arbitrary Versapay URIs
This is extracted from an application that had a requirement to build
some arbitrary Versapay URIs. Hopefully this method can be made more
useful with time to build more specific end point helpers.

Note that we're assuming Rails is used by the end user with this method.
I felt that this was safe given the implicit requirement of Rails in
other methods.